kreirana tabela customers_tico

+++ database/migrations/2021_01_03_184323_create_articles_table.php
@@ -13,27 +13,29 @@ class CreateArticlesTable extends Migration
      */
     public function up()
     {
-        Schema::create('articles', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->integer('code')->unique();
-            $table->string('title');
-            $table->unsignedBigInteger('categories_id');
-            $table->foreign('categories_id')->references('id')->on('categories')->onUpdate('cascade')->onDelete('cascade');
-            $table->integer('price');
-            $table->integer('amount');
-            $table->integer('discount')->default(0);/*
-            $table->integer('discount_silver')->default(0);
-            $table->integer('discount_gold')->default(0);
-            $table->integer('discount_premium')->default(0);*/
-            $table->string('image')->nullable();
-            $table->string('brand')->nullable();
-            $table->integer('rating')->nullable();
-            $table->text('tags')->nullable();
-            $table->longText('description')->nullable();
-            $table->text('short_description')->nullable();
-            $table->timestamp('created_at')->useCurrent();
-            $table->timestamp('updated_at')->useCurrent();
-        });
+        if(!Schema::hasTable('articles')){
+            Schema::create('articles', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->integer('code')->unique();
+                $table->string('title');
+                $table->unsignedBigInteger('categories_id');
+                $table->foreign('categories_id')->references('id')->on('categories')->onUpdate('cascade')->onDelete('cascade');
+                $table->integer('price');
+                $table->integer('amount');
+                $table->integer('discount')->default(0);/*
+                $table->integer('discount_silver')->default(0);
+                $table->integer('discount_gold')->default(0);
+                $table->integer('discount_premium')->default(0);*/
+                $table->string('image')->nullable();
+                $table->string('brand')->nullable();
+                $table->integer('rating')->nullable();
+                $table->text('tags')->nullable();
+                $table->longText('description')->nullable();
+                $table->text('short_description')->nullable();
+                $table->timestamp('created_at')->useCurrent();
+                $table->timestamp('updated_at')->useCurrent();
+            });
+        }
     }
 
     /**
